Rename repository package
