+++ repository/axiom/reservations.go
@@ -1,4 +1,4 @@
-package mongo
+package axiom
 
 import (
 	"context"
